Move painting code into paint module
